+++ Controllers/ProductoController.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using TiendaVirtualYanten.Data;
+using TiendaVirtualYanten.Helpers;
 using TiendaVirtualYanten.Models;
 
 namespace TiendaVirtualYanten.Controllers
@@ -8,14 +9,22 @@ namespace TiendaVirtualYanten.Controllers
     public class ProductoController : Controller
     {
         private readonly TiendaContext _context;
+        private readonly PermisoHelper _permisos;
 
-        public ProductoController(TiendaContext context)
+        public ProductoController(TiendaContext context, PermisoHelper permisos)
         {
             _context = context;
+            _permisos = permisos;
         }
 
         public IActionResult Index()
         {
+            if (!_permisos.Tiene("Producto", "Ver"))
+                return RedirectToAction("Denegado", "Home");
+
+            ViewBag.PuedeCrear = _permisos.Tiene("Producto", "Crear");
+            ViewBag.PuedeEditar = _permisos.Tiene("Producto", "Editar");
+            ViewBag.PuedeEliminar = _permisos.Tiene("Producto", "Eliminar");
             var productos = _context.Productos
                 .Include(p => p.Categoria)
                 .ToList();
@@ -24,6 +33,9 @@ namespace TiendaVirtualYanten.Controllers
 
         public IActionResult Create()
         {
+            if (!_permisos.Tiene("Producto", "Crear"))
+                return RedirectToAction("Denegado", "Home");
+
             ViewBag.Categorias = _context.Categorias.ToList();
             return View();
         }
@@ -31,6 +43,9 @@ namespace TiendaVirtualYanten.Controllers
         [HttpPost]
         public IActionResult Create(Producto producto)
         {
+            if (!_permisos.Tiene("Producto", "Crear"))
+                return RedirectToAction("Denegado", "Home");
+
             if (ModelState.IsValid)
             {
                 _context.Productos.Add(producto);
@@ -43,6 +58,9 @@ namespace TiendaVirtualYanten.Controllers
 
         public IActionResult Edit(int id)
         {
+            if (!_permisos.Tiene("Producto", "Editar"))
+                return RedirectToAction("Denegado", "Home");
+
             var producto = _context.Productos.Find(id);
             if (producto == null) return RedirectToAction("Index");
             ViewBag.Categorias = _context.Categorias.ToList();
@@ -52,6 +70,9 @@ namespace TiendaVirtualYanten.Controllers
         [HttpPost]
         public IActionResult Edit(Producto producto)
         {
+            if (!_permisos.Tiene("Producto", "Editar"))
+                return RedirectToAction("Denegado", "Home");
+
             if (ModelState.IsValid)
             {
                 _context.Productos.Update(producto);
@@ -64,6 +85,9 @@ namespace TiendaVirtualYanten.Controllers
 
         public IActionResult Delete(int id)
         {
+            if (!_permisos.Tiene("Producto", "Eliminar"))
+                return RedirectToAction("Denegado", "Home");
+
             var producto = _context.Productos.Find(id);
             if (producto == null) return RedirectToAction("Index");
             _context.Productos.Remove(producto);
